Update DroidScript to allow OTG connection
To open up connection methods so that Bluetooth is not necessary. This reduces power consumption and allows for the potential of a battery powered device.
There is also the possibility that the TUC can run on kinetic energy. Power provided from the OTG cable would allow connection when machine is not running